requirements.txt
python-dotenv~=1.0.1
aiogram~=3.3.0
aiohttp~=3.9.3